New class BattleSystem, Transfer some part from BattleScene to BattleSystem, Adding of SpriteAnimation h and cpp.

+++ Base/Source/Main/Game/Miscellaneous/RouteGeneration.cpp
@@ -15,7 +15,7 @@ std::vector<TerrainNode*> RouteGeneration::GeneratePath()
 
 	DefinedPath.push_back(StartNode);
 
-	while (DefinedPath.size() < MaximumPathCount)
+	while (DefinedPath.size() < (unsigned)MaximumPathCount)
 	{
 		float ClosestDistance = FLT_MAX;
 
@@ -32,7 +32,7 @@ std::vector<TerrainNode*> RouteGeneration::GeneratePath()
 			ClosestDistance = MostViableNode->CalculateRemainderPathCost(TargetNode);
 		}
 		// A distance comparison between each node, in an attempt to find the closest.
-		for (int i = 0; i < PossibleMoves.size(); ++i)
+		for (unsigned int i = 0; i < PossibleMoves.size(); ++i)
 		{
 			if (PossibleMoves[i] != PreviousNode)
 			{
